feat: wire list to ff component and move it to its own folder
