+++ .changeset/gold-eggs-pump.md
@@ -1,0 +1,6 @@
+---
+"@devtools/feature-flags": minor
+"@ledgerhq/web-tools": minor
+---
+
+Add the list of flags in feature-flags devtool. Add a toggle to enable/disable the feature-flags. And link it to web-tools
