+++ devtools/feature-flags/src/feature-flags/FeatureFlags.tsx
@@ -1,0 +1,10 @@
+import { FlagList } from "../components";
+import type { FeatureFlagsToolProps } from "../types";
+
+export const FeatureFlags = (props: FeatureFlagsToolProps) => (
+  <div>
+    <FlagList {...props} />
+  </div>
+);
+
+export default FeatureFlags;
