+++ devtools/feature-flags/src/index.ts
@@ -1,4 +1,4 @@
-import FeatureFlags from "./FeatureFlags";
+import FeatureFlags from "./feature-flags/FeatureFlags";
 
 export type { FeatureFlagsToolProps, FlagDisplayState, FlagFilter } from "./types";
 export { ALL_FLAG_IDS } from "./constants";
